add addressing types / load
NOTE: Nothing is tested yet. I am unsure if the page crossing is for every instruction. Some documents don't even mention the extra cycle for page crossing, some do but only for certian instructions, some mention it for all instructions.

I've added a bool flag for now so i can easily on/off the extra cycle. Not that the 1 cycle is that important anyway.