improve clock cycle accuracy
not perfect yet as its out of sync by 10 ticks after 5000 op calls...which is pretty good for a first test!

The tick imp isn't great either, in fact, its very bad.

There's a pattern to when a cycle happens, need to explore it. also when writing to mem in some(?) ops, the nes actually does a double write. Also in some cases it does a fake read (called a dummy read). These don't actually effect data at all, but this is why my cycles are probably out by 10.