Added ppu read / write for registers
Added support for this in the mmu.

NOTE: The mmu will be removed (i think) and just have it in the cpu.c. This will mean that I can inline the function, which is very import because it is often called ~50,000+ a second.

Also it would make more sense since the the addresses are actually just a cpu memory map.