Add cart.c/.h for rom loading and header parsing.
Header parsing isn't done yet, but rom loading works.

Added check to ensure that the rom is no larger than 1MiB.

Added error messages for reset functions.

All errors are sent to stderr now.

quickloading the rom is added, so you can pass the rom as an arg and it'll load it.