Add basic mapper groundwork
Had a quick idea on how to implement the mapper(s). I've made a note in the mapper.c file on the 2 ideas that I thought of (right before I was falling asleep!)

I've made it so that by loading the rom in cart.c, it also loads the mapper. This makes the most sense as i need to configure the mapper based on the rom size. Such as, mapper 0 is either 16kb or 32kb. If 16kb, I need to mirror reads. Else 32kb, I can read from the full address.

I also reset the mapper inside cart reset.

I disabled -O3 optimise to improve compile times by like 3 seconds ;) And seeing as I don't need to omptimise anything atm.